lab/gcp/22-07-2025-bollette/func_stat/requirements.txt
google-cloud-firestore==2.22.0

python-dateutil==2.9.0


# Se contiene una HTTP function 
flask==2.3.3